refactor: move helper
